# contracts/api.yaml
...
openapi: 3.1.0
info:
  title: FinDesk API
  version: 1.0.0
  description: >
    Control plane for the autonomous CFO. All endpoints are tenant-scoped via
    the access token's active-tenant claim. Money is integer paise; timestamps
    UTC ISO-8601; IDs UUIDv7.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /approvals/{approval_id}/decide:
    parameters: [{ name: approval_id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [approvals]
      summary: Approve/reject; on approve mints single-use action token and resumes the run
      responses: { "200": { description: decided } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }